+++ paddle/fluid/framework/ir/layer_norm_fuse_pass.cc
@@ -378,7 +378,8 @@ void LayerNormFusePass::ApplyImpl(Graph* graph) const {
     setIntermediateOut(&ln_op_desc, "Mean", scope_name_);
     setIntermediateOut(&ln_op_desc, "Variance", scope_name_);
     ln_op_desc.SetAttr("begin_norm_axis", begin_norm_axis);
-    ln_op_desc.SetAttr("epsilon", *(eps_tensor->data<float>()));
+    ln_op_desc.SetAttr("epsilon",
+                       static_cast<double>(*(eps_tensor->data<float>())));
     ln_op_desc.SetAttr("is_test", true);
 
     if (!IsCompat(ln_op_desc)) {
